requirements.txt
supabase==2.3.4
schedule==1.2.0
python-dotenv==1.0.0
